requirements.txt
python-telegram-bot==21.7
openai==1.51.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pillow==11.0.0
pytz==2024.2
python-dotenv==1.0.1
aiofiles==24.1.0
redis==5.0.1
